feat: chapter 4 - persisting ui state
