+++ app/src/main/java/com/bignerdranch/android/tbnr_geoquiz/QuizViewModel.kt
@@ -1,0 +1,32 @@
+package com.bignerdranch.android.tbnr_geoquiz
+
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+
+const val CURRENT_INDEX_KEY = "CURRENT_INDEX_KEY"
+
+class QuizViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel() {
+
+    private val questionBank = listOf(
+        Question(R.string.question_australia, true),
+        Question(R.string.question_oceans, true),
+        Question(R.string.question_mideast, false),
+        Question(R.string.question_africa, false),
+        Question(R.string.question_americas, true),
+        Question(R.string.question_asia, true)
+    )
+
+    private var currentIndex: Int
+        get() = savedStateHandle[CURRENT_INDEX_KEY] ?: 0
+        set(value) = savedStateHandle.set(CURRENT_INDEX_KEY, value)
+
+    val currentQuestionAnswer: Boolean
+        get() = questionBank[currentIndex].answer
+
+    val currentQuestionText: Int
+        get() = questionBank[currentIndex].textResId
+
+    fun moveToNext() {
+        currentIndex = (currentIndex + 1) % questionBank.size
+    }
+}
